requirements.txt
djangorestframework==3.14.0
drf-yasg==1.21.4